requirements.txt
aiogram~=3.18.0
asyncio
python-dotenv~=1.0.1